+++ lib/kitchen/directions/bake_chapter_introductions.rb
@@ -3,21 +3,31 @@
 module Kitchen
   module Directions
     module BakeChapterIntroductions
-      def self.v1(book:)
+      def self.v1(book:, bake_chapter_objectives: true)
         book.chapters.each do |chapter|
-          outline_items_html = chapter.pages.map do |page|
-            next if page.is_introduction?
+          chapter_outline_html = ''
 
-            <<~HTML
-              <div class="os-chapter-objective">
-                <a class="os-chapter-objective" href="##{page.title[:id]}">
-                  <span class="os-number">#{chapter.count_in(:book)}.#{page.count_in(:chapter) - 1}</span>
-                  <span class="os-divider"> </span>
-                  <span data-type="" itemprop="" class="os-text">#{page.title.children[0].text}</span>
-                </a>
+          if bake_chapter_objectives
+            outline_items_html = chapter.non_introduction_pages.map do |page|
+
+              <<~HTML
+                <div class="os-chapter-objective">
+                  <a class="os-chapter-objective" href="##{page.title[:id]}">
+                    <span class="os-number">#{chapter.count_in(:book)}.#{page.count_in(:chapter)}</span>
+                    <span class="os-divider"> </span>
+                    <span data-type="" itemprop="" class="os-text">#{page.title.children[0].text}</span>
+                  </a>
+                </div>
+              HTML
+            end.join('')
+
+            chapter_outline_html = <<~HTML
+              <div class="os-chapter-outline">
+                <h3 class="os-title">#{I18n.t(:chapter_outline)}</h3>
+                #{outline_items_html}
               </div>
             HTML
-          end.join('')
+          end
 
           introduction_page = chapter.introduction_page
 
@@ -33,10 +43,7 @@ module Kitchen
           introduction_page.append(child:
             <<~HTML
               <div class="intro-body">
-                <div class="os-chapter-outline">
-                  <h3 class="os-title">#{I18n.t(:chapter_outline)}</h3>
-                  #{outline_items_html}
-                </div>
+                #{chapter_outline_html}
                 <div class="intro-text">
                   #{title.paste}
                   #{intro_content.paste}