+++ spec/directions/bake_chapter_introductions_spec.rb
@@ -46,7 +46,7 @@ RSpec.describe Kitchen::Directions::BakeChapterIntroductions do
     )
   end
 
-  context 'when v1 called on book' do
+  context 'when v1 called on book with chapter objectives' do
     it 'works' do
       described_class.v1(book: book)
       expect(book.body).to match_normalized_html(
@@ -128,4 +128,55 @@ RSpec.describe Kitchen::Directions::BakeChapterIntroductions do
       }.from("./*[@data-type = 'document-title']").to(".intro-text > [data-type='document-title']")
     end
   end
+
+  context 'when v1 called on book without chapter objectives' do
+    it 'works' do
+      described_class.v1(book: book, bake_chapter_objectives: false)
+      expect(book.body).to match_normalized_html(
+        <<~HTML
+          <body>
+            <div data-type="chapter">
+              <h1 data-type="document-title">Chapter 1 Title</h1>
+              <div class="introduction" data-type="page">
+                <div data-type="metadata">don't touch this</div>
+                <figure class="splash">can't touch this (stop! hammer time)</figure>
+                <div class="intro-body">
+                  <div class="intro-text">
+                    <h2 data-type="document-title">
+                      <span class="os-text" data-type="" itemprop="">Introduction 1</span>
+                    </h2>
+                    <figure>move this</figure>
+                    <div>content</div>
+                  </div>
+                </div>
+              </div>
+              <div data-type="page">
+                <div data-type="document-title" id="becomes-ref-link">should be objective 1.1</div>
+              </div>
+              <div data-type="page">
+                <div data-type="document-title">should be objective 1.2</div>
+              </div>
+            </div>
+            <div data-type="chapter">
+              <h1 data-type="document-title">Chapter 2 Title</h1>
+              <div class="introduction" data-type="page">
+                <div class="intro-body">
+                  <div class="intro-text">
+                    <h2 data-type="document-title">
+                      <span class="os-text" data-type="" itemprop="">Introduction 2</span>
+                    </h2>
+                    <div>content</div>
+                  </div>
+                </div>
+              </div>
+              <div data-type="page">
+                <div data-type="document-title">should be objective 2.1</div>
+              </div>
+            </div>
+          </body>
+        HTML
+      )
+    end
+  end
+
 end